requirements.txt
# ── Servidor Web ────────────────────────────────────────────────────────
Flask==3.0.3
gunicorn==22.0.0

# ── Banco de dados ──────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36

# ── Segurança ───────────────────────────────────────────────────────────
Werkzeug==3.0.4

# ── Variáveis de ambiente ───────────────────────────────────────────────
python-dotenv==1.0.1
